+++ xair_api/adapter.py
@@ -1,4 +1,5 @@
 from .bus import Bus as IBus
+from .headamp import HeadAmp as IHeadAmp
 from .lr import LR as ILR
 from .rtn import AuxRtn as IAuxRtn
 from .rtn import FxRtn as IFxRtn
@@ -38,3 +39,9 @@ class Matrix(ILR):
     @property
     def address(self) -> str:
         return f"/mtx/{str(self.index).zfill(2)}"
+
+
+class HeadAmp(IHeadAmp):
+    @property
+    def address(self):
+        return f"/headamp/{str(self.index).zfill(3)}"